Tweak how firewall MAC add works (see body)
When running an apply via flight profile for multiple nodes, the playbook
ends up getting a bit confused with all the add MAC -> reload loops
going at once. To circurmvent this I've tweaked it to add all MACs
permanently to firewall then do a reload. This seems to reduce the
likelihood of an error occuring.